chore: remove obsolete utils warnings
